device_app/pyproject.toml
[project]
name = "camera_app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.7",
    "async-timeout>=5.0.1",
    "pillow==10.4.0",
    "pydoover",
]

[project.scripts]
doover-app-run = "camera_app:main"
export-config = "camera_app.app_config:export"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "aiohttp>=3.11.18",
    "grpcio>=1.71.0",
    "protobuf>=6.30.2",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "requests>=2.32.3",
]

[tool.uv.sources]
pydoover = { git = "https://github.com/getdoover/pydoover" }
